+++ price-range/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log
 
+## [1.2.0] (2020.04.06):
+### Added
+- Ability to enter only the minimum or maximum price value
+### Fixed
+- Price range issue related to product discount and minimal quantity
+
 ## [1.1.0] (2020.04.04):
 ### Added
 - Compatibility with the Live Options+ module